+++ feature-staking-impl/src/main/java/jp/co/soramitsu/feature_staking_impl/presentation/validators/current/CurrentValidatorsAdapter.kt
@@ -10,6 +10,8 @@ import jp.co.soramitsu.common.utils.setCompoundDrawableTint
 import jp.co.soramitsu.common.utils.setTextColorRes
 import jp.co.soramitsu.common.utils.setTextOrHide
 import jp.co.soramitsu.feature_staking_impl.R
+import jp.co.soramitsu.feature_staking_impl.presentation.validators.current.model.NominatedValidatorModel
+import jp.co.soramitsu.feature_staking_impl.presentation.validators.current.model.NominatedValidatorStatusModel
 import kotlinx.android.synthetic.main.item_current_validator.view.itemCurrentValidatorIcon
 import kotlinx.android.synthetic.main.item_current_validator.view.itemCurrentValidatorName
 import kotlinx.android.synthetic.main.item_current_validator.view.itemCurrentValidatorNominated
@@ -17,6 +19,7 @@ import kotlinx.android.synthetic.main.item_current_validator_group.view.itemCurr
 import kotlinx.android.synthetic.main.item_current_validator_group.view.itemCurrentValidatorGroupStatus
 
 class CurrentValidatorsAdapter : GroupedListAdapter<NominatedValidatorStatusModel, NominatedValidatorModel>(CurrentValidatorsDiffCallback) {
+
     override fun createGroupViewHolder(parent: ViewGroup): GroupedListHolder {
         return CurrentValidatorsGroupHolder(parent.inflateChild(R.layout.item_current_validator_group))
     }
